backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
pydantic>=2.10.0
reportlab==4.2.5
openpyxl==3.1.5

